// src/models/common/thunkStatus.enum.ts
export enum ThunkStatusEnum {
  IDLE = 'idle',
  LOADED = 'loaded',
  LOADING = 'loading',
}

export enum ThunkResultEnum {
  FAILURE = 'failure',
  FAILURE_OPERATION_DELETED_ENTITY = 'already_deleted',
  SUCCESS = 'success',
}
export enum LeadListEnum {
  ALL_LEADS = 'ALL_LEADS',
  SEARCHED_LEADS = 'SEARCHED_LEADS',
  FILTERED_LEADS = 'FILTERED_LEADS',
}